requirements.txt
pandas==0.23.4
xlrd==1.2.0
python-docx==0.8.7
